requirements.txt
django==1.9
djangorestframework
markdown       # Markdown support for the browsable API.
django-filter  # Filtering support

